+++ src/shared/containers/index.tsx
@@ -69,7 +69,7 @@ export const RouteToDashboardList = lazy(() =>
 )
 export const FlowsIndex = lazy(() => import('src/flows/components/FlowsIndex'))
 export const NotFound = lazy(() => import('src/shared/components/NotFound'))
-export const UsersPage = lazy(() => import('src/users/components/users/Users'))
+export const UsersPage = lazy(() => import('src/users/components/Users'))
 export const UsagePage = lazy(() => import('src/usage/UsagePage'))
 export const BillingPage = lazy(() =>
   import('src/billing/components/BillingPage')
