+++ src/style/chronograf.scss
@@ -126,7 +126,7 @@
 @import 'src/cloud/components/RateLimitOverlay.scss';
 @import 'src/cloud/components/experiments/DatalessEmptyState.scss';
 @import 'src/shared/components/notifications/Notification.scss';
-@import 'src/users/components/users/UserList.scss';
+@import 'src/users/components/UserList.scss';
 @import 'src/billing/components/PayAsYouGo/cancellation-overlay.scss';
 @import 'src/billing/components/PayAsYouGo/PlanTypePanel.scss';
 @import 'src/billing/components/BillingPage.scss';
