+++ src/users/components/Users.tsx
@@ -3,7 +3,7 @@ import React, {FC} from 'react'
 
 // Components
 import UsersProvider from 'src/users/context/users'
-import UsersPage from 'src/users/components/users/UsersPage'
+import UsersPage from 'src/users/components/UsersPage'
 
 const Users: FC = () => (
   <UsersProvider>
